+++ client/tasks/index.ts
@@ -2,7 +2,7 @@
  * Internal dependencies
  */
 import { Tasks } from './tasks';
-import './tasks-fills';
+import './fills';
 
 export * from './placeholder';
 export default Tasks;
